requirements.txt
selenium
undetected_chromedriver
google-cloud-pubsub
avro
